Remove custom editor module
