+++ README.md
@@ -1,10 +1,10 @@
 
-This project contains a minimal note-taking application written with [Textual](https://textual.textualize.io/). Notes are edited using a custom `NoteEditor` built on top of *prompt_toolkit* which provides soft wrapping and clipboard shortcuts. An optional countdown timer can help manage your time.
+This project contains a minimal note-taking application written with [Textual](https://textual.textualize.io/). Notes are edited using Textual's built-in `TextArea` widget with soft wrapping and clipboard shortcuts. An optional countdown timer can help manage your time.
 All interface labels and notifications are in Danish for localization.
 
 ## Usage
 
-1. Install dependencies with `pip install textual prompt_toolkit`.
+1. Install dependencies with `pip install textual`.
 2. From the project folder run the app using `python main.py`.
 3. Create additional tabs with `Ctrl+N` or open an existing file with `Ctrl+O`.
    Close the active tab with `Ctrl+W` and hide/show the tab bar with `Ctrl+B`.
